+++ QuickStackStore/QuickStackStorePlugin.cs
@@ -12,9 +12,10 @@ namespace QuickStackStore
     {
         public const string GUID = "goldenrevolver.quick_stack_store";
         public const string NAME = "Quick Stack - Store - Sort - Trash - Restock";
-        public const string VERSION = "1.3.3";
+        public const string VERSION = "1.3.4";
 
-        protected void Awake()
+        // intentionally not Awake, so the chainloader is done (for compatibility checks, mostly in the config)
+        protected void Start()
         {
             if (CompatibilitySupport.HasOutdatedMUCPlugin())
             {